chore(scan): require layouts in ballot packages
Closes #1595
Refs #1524

We've been building ballot package with pairs of PDFs and their associated layout JSON files for a while, but we would fall back on re-interpreting the PDFs in the event that a ballot package being loaded didn't have the layout information.

We now enforce that the layout files are present, which means that our old ballot packages from before #1596 or so will no longer work. I updated the ballot packages present in the repo fixtures to include layouts if they didn't already.