feat(res-to-ts): add Election representations
When we can tell that a JSON file is an election definition, generate code to export `electionDefinition` and `election` since that's the most likely thing any users of those fixtures will want to do.

Also adds the SHA-256 hash of the file to all the comments for exports to make it easier to find things by hash and know what its hash is, especially election definitions.