refactor(fixtures): use Election exports
Now that we have `election` and `electionDefinition` exports for free, let's use them. This commit also takes all the `WithDataFiles` objects and makes them be an `export * as …` from `index.ts` files that now handle bundling all the relevant fixture files together into a neat and tidy package.

This commit removes the ".original" suffix we had on a bunch of the fixture data files since we're no longer building the `.ts` files alongside their original counterparts.