feat(fixtures): add images of stretched ballot
This ballot comes from VxScan where we've been seeing spurious overvotes (#1959). The overvotes are likely due to the same issue that #2053 addressed for NH ballots. Since our ballots do not use timing marks we won't be able to fix it in exactly the same way. This commit adds the fixture images for a future commit that will build a test around them.