fix(vx-scan): reject overly stretched ballots
Fixes #1959

The precinct scanner can sometimes stretch one side of the ballot in such a way that the ballot images "lean" in one direction or the other. This causes spurious marks to be detected, which usually ends up causing spurious overvotes. We may be able to fix this in the future, but for now we just reject anything that is obviously stretched.