fix(vx-scan): better account for stretching
This change reverts the change that rejected ballots that were too stretched. Instead, we simply allow the "jiggling" to find the likely target to move up to 5px instead of just 2px. This was enough for the few tests I did, but is hopefully conservative enough not to introduce a new failure mode.

This is a more proper fix for #1959.