fix(scan): Ignore extraneous commands to precinct scanner state machine
This prevents race conditions when the frontend has an outdated scanner status and tries to send a command. E.g. if the frontend thinks the scanner is in `ready_to_scan` state but the paper got quickly removed and it's back to `no_paper`, it might still try to send a `SCAN` command.

This also allows the frontend to not have to worry about sending commands multiple times, which makes the frontend logic much simpler.