fix(scan): Interpret blank BMD pages correctly for precinct scanner (#2220)
There was a special case to catch blank pages of a BMD ballot (i.e. the back of the ballot) in the interpret_vx worker that wasn't included in the new simple interpreter used by the precinct scanner. I moved the logic into the interpreter itself to make sure both central and precinct scanners can use the same logic.