refactor(scan): Combine page interpretations in SimpleInterpreter
Instead of combining the interpretations for the two pages of a ballot in the state machine module, we do it behind the interface of the interpreter. This will allow easier mocking of the interpreter in tests.

I also changed the types so that the state machine event type isn't reused as the interpretation result type. Instead, I created a new type to represent the interpretation of a ballot sheet.