fix(scan): Handle insertion of second ballot while first ballot is accepting
If a second ballot is inserted while the first ballot is in the middle of being accepted (a very short window, but possible), the first ballot is successfully dropped but plustek returns a jam error. By ignoring that error, we can correctly record that the ballot got accepted.

Then, we also add a delay to make sure that the user has time to see the "ballot counted" screen before the second ballot starts scanning.