refactor(scan): Make the calibrate API blocking
I had changed the calibrate API to be non-blocking (to be similar to the rest of the scanner commands). However, due to the way that calibrate is actually used in the app (on the admin screen only), I think it makes more sense to have a blocking API call. That way if the admin screen is exited by card removal in the middle of calibration (or after), the frontend doesn't need to make an API request to exit calibration, it will automatically go back to no_paper when its done. This seems less prone to getting stuck in a calibration state if the frontend messes up.