refactor(scan): Move accepted timeout to the scan service
Previously, we had let the frontend decide when to timeout the "accepted" state by hitting the `/wait-for-paper` endpoint. Simplify by moving this to the scan service as a configurable param to the state machine. (This also preps for adding another delay to the accepted state for when to start scanning newly inserted ballots)