fix(scan): Fix bad state reference
It turns out that you can't do relative state names (e.g.
`..ready_to_scan`). Usually I get a type error when a state name is invalid, but for some reason this didn't give me one so I thought it worked. Instead, in order to reference a state from a child state, you have to use the `id` property to give it a global name, and then reference it with a `#` (e.g. `#ready_to_scan`).