feat(scan): Hold paper in front on reboot (#2288)
If there's a paper held in the front on startup (e.g. after returning or rejecting a ballot before rebooting), then we shouldn't reject it (which will spit it out completely), but rather just hold it where it is (already rejected). This is consistent with our "always hold papers in front" approach.