fix(ps): Don't crash if card removed during calibration
Previously, we asserted that the scanner would never be in the `calibrating` state while in voter mode. That's not true - if the election manager removes their card during calibration and the polls are open, it's possible. Since this isn't something we really expect to happen, we just handle it by rendering a blank screen (which will go away when calibration finishes). If we decide we need a screen here, we can come back and add it later. The goal of this PR is to not crash.