feat(scan): Kill unresponsive plustekctl
In some cases, when `plustekctl` becomes unresponsive, it will not respond to the `quit` command that we try to send to disconnect. That will block a new `plustekctl` process from connecting to the scanner (seemingly, the old process is still holding some lock).

In this branch, we add a new command `kill` to the plustek-sdk library to send a `SIGTERM` to `plustekctl`. When we try disconnecting after an unexpected error or timeout, we wait a bit and then give up and send the kill signal.

This branch also makes a few changes to the frontend:
- When we're recovering from an error, show the scanning progress screen. More often than not, these types of errors occur during scanning, and continuing to show that screen makes for a seamless UX. We could probably get fancier and find a way to keep the state machine on whatever state its on when the error occurs, but that would be a bigger change.
- Make sure actual unrecoverable errors show a message to restart the scanner, with simplified copy from @mattroe.