fix(scan): Simplify debug logging
We don't want production logs and debug dev logs showing at the same time in dev, it's TMI. Instead, we use the built in option of the logger to use a debugger instead of logging in dev.

However, in order to get the simple debug messages we had before, I needed to wrap the debugger in a function to take the log line and only pull out the relevant message field.