fix(ps): Don't show scanning screen on accept after review (#2354)
Accepting a ballot after review only takes about half a second. Previously, we showed the "Please wait, checking your ballot" screen, since it used the same code path as a ballot that didn't need review.

To fix this, we add a new state `accepting_after_review` to the state machine (which uses the same implementation as `accepting` - finally figured out how to reuse state machine logic).

The frontend is updated to disable the buttons in the modal for that half second, which provides enough visual feedback that the button press actually did something without being too disruptive.