refactor(scan): Store configuration in interpreter, not state machine
Previously, the state machine knew about interpreter configuration. Since the state machine is already pretty complicated and the configuration is specifically for the interpreter, here we move caching of the config to the interpreter. (It's "caching" because the store has the actual source of truth for the config, but it's slow enough to read that we cache it in memory and reload on writes.)

Plus, the scanner can technically still function without configuration (specifically for calibration), so if we ever wanted to allow that in the frontend, it makes sense to decouple that from the state machine.