fix(scan): Fail when Plustek only returns one scanned page
Previously, if Plustek only returned one page, the scanner would hang (due to unclear hanging behavior from canvas.loadImage). One contributing factor is that we were converting an array of files returned by plustekctl into a pair of files without checking that the array had exactly two files in it. While we don't know the root cause of why plustekctl only returns one file sometimes, we can at least fail better, rejecting the ballot so it can be scanned again.

Here, we modify plustek-sdk to validate that two files were returned, and fail otherwise. The precinct scanner state machine also needed to be modified to better handle scanning errors, since this particular type of error results in the paper being kept in the back. We hadn't seen this type of scanning error before, so we weren't handling it well (going straight to the `rejected` state, which assumes the ballot is still in the front). Now, we check where the paper is before deciding what state to go to.