fix(ps): Detect Plustek disconnection with scan service, not kiosk browser
Currently, we have two ways of telling if the Plustek scanner is
disconnected: kiosk browser's USB device list and the scan service
scanner status. Kiosk browser can't differentiate between a powered off/unplugged scanner and a crashed scanner, while the scan service can. Here we change the frontend to only check the scan service status and ignore kiosk browser in order to have more precise error messages when the Plustek is not reachable.