build(frontends): make `build` do both `tsc` and `vite`
While the production build using `vite` does not require any of the libraries to be built in place, the prodserver requires `@votingworks/logging` to be built or starting BMD will fail. There is a different long-term fix for this which will transition us to a more "no builds in dev" and "bundle all the things in prod" system, but for now an easy workaround is just to ensure that both styles of builds are done: `tsc` for in-place building and `vite` for production bundling.