fix(scan): Reconnect to plustek after any unexpected scanning error
Previously, if we got an unexpected scanning error and the Plustek was still responsive, we would either automatically retry scanning or reject and let the voter retry, depending on the paper status. Now, if we get an unexpected scanning error, we handle it like any other unexpected error, reconnecting to Plustek. This will hopefully make us more resilient to scanning errors where we don't know the root cause.