chore(scan): remove write-in area scoring
We don't have a customer that needs this right now and it makes #2459 more complicated because it requires us to have the ballot template images when interpreting. If we want to bring this back I suggest we encode the area to look for a write-in mark within the layout of the ballots and assume the entire area is white and/or make the threshold configurable to account for the inclusion of some black lines or text.