fix(scan): Timeout for accept
If the accept command returns successfully, it may or may not actually be done accepting the ballot (e.g. if there is an obstruction holding the ballot back). In order to make sure the ballot was actually dropped, we check the paper status. Since the rollers may still be trying to drop the ballot, we need to wait a few seconds to see if the ballot gets dropped or not. If it doesn't then we're safe to reject. Previously, we tried rejecting too early which led to a series of confusing paper statuses.