fix(scan): Require restart when only one scan image file returned (#2472)
When Plustek only returns one scanned image file, it seems to be
unrecoverable even if we disconnect/reconnect to Plustek. Subsequent scans exhibit the same buggy behavior.

Restarting the tablet seems to fix it, so we make the state machine go to the unrecoverable_error state, which shows the user instructions to restart.