refactor(admin): introduce `backend` abstraction (#2501)
This is a continuation of #2477. That PR added `react-query` and `useElectionManagerStore` to encapsulate it. Previously, `useElectionManagerStore` took a `storage` prop to control how the data was stored and retrieved to make testing and encapsulation easier. This PR replaces `storage` with `backend` and introduces a storage-based backend that it still uses for now, but will be replaced with an admin service-based `backend`. Additionally, a `memory`-based `backend` is introduced to make testing easier.