test(scan): fix `act` warning
This test has async operations that happen after the test has been torn down, leading to a warning about using `act`. This commit adds an `await` to find something on the screen after the action has been performed to ensure the async stuff all happens by the time the test is done.