test(scan): Fix flaky precinct scanner test (#2515)
This test was flaky because sometimes the precinct scanner state machine would get to the `returned` state before the second ballot was inserted. To prevent the chance of that happening, I removed a check for `returning` state that delayed insertion of the second ballot. It's not an important check since it's covered by other tests.