feat(fixtures): add ballot images & layout to CVRs
Adds a `--includeBallotImages` flag to `generate-sample-cvr-file` CLI. To do this in a way that had some fidelity to it, I had to switch from using an election definition to a ballot package. This way we can generate images for the CVRs and get the real layout information. Because of the complexity of generating CVRs for multi-sheet ballots, I opted to assert that this only works with single-sheet ballots for now.