feat(fixtures): allow generating CVRs with ballot images (#2516)
* refactor(fixtures): extract CVR generator

Extracts the `generate_cvrs.ts` functionality to its own package called `cvr-fixture-generator`. This will help prevent dependency cycles as `fixtures` depends on things, but a lot of packages depend on `fixtures` for testing purposes.

* feat(fixtures): improve CVR generator

- programmatic API takes an election definition rather than an election
- CLI writes to stdout by default rather than `output.jsonl`
- CLI waits until the stream has finished writing to return
- CLI now has tests

* feat(fixtures): add ballot package for "minimal exhaustive"

I plan to use the ballot package as the source file for generating sample CVRs with ballot images, so this election definition is a useful one to have a ballot package for.

* feat(image-utils): add image encoding functions

These functions allow taking an `ImageData` and encoding it as PNG or JPEG for use in a stream, or as a data URL.

* feat(utils): add `mapAsync` and `takeAsync`

These are useful when dealing with async generators.

* feat(fixtures): add ballot images & layout to CVRs

Adds a `--includeBallotImages` flag to `generate-sample-cvr-file` CLI. To do this in a way that had some fidelity to it, I had to switch from using an election definition to a ballot package. This way we can generate images for the CVRs and get the real layout information. Because of the complexity of generating CVRs for multi-sheet ballots, I opted to assert that this only works with single-sheet ballots for now.

* chore(types): fix `codemods` types

Something in a previous commit jumbled the dependencies around and now the babel deps are out of sync. This should fix it.

* fix(scan): pass only the page layouts present in the CVR