refactor(admin): make store updates more granular
When making changes to CVRs, i.e. loading a new one or clearing the list, we previously would get the list of all the CVRs and build a new one from it, then set the whole thing again. Instead, we now have distinct operations for the actions we want to perform: `add` and `clear`.

This same thing applies to the external tallies, with the addition that we now enforce that there is only ever at most one of each external tally source type.

The next step here is to add API endpoints for the CVRs in the admin service corresponding to the `add` and `clear` actions.