refactor(admin): store CVRs in `admin` backend
Adds endpoints to import CVR files and process them. Adds a join table for `cvr_files` and `cvrs` so that duplicate CVRs are linked to both files. Changes the frontend to use the new endpoints for adding and clearing files, but also preserves the existing storage for now to keep the tally code the same.