build(node): install NodeJS from nodejs.org
There are differences among the 16.x versions of NodeJS that are relevant to our development. For example, `Intl.DateTimeFormat` yields slightly different results between 16.14.2 and 16.17.0. We can't easily install a specific version from nodesource [1], so instead this commit changes the `setup-dev` script to just download the right tarball and unpack it into `/usr/local`.

[1]: https://github.com/nodesource/distributions/issues/1427#issuecomment-1207640371