feat(admin): make write-in transcription work (#2558)
* feat(utils): add collections helpers

Adds `map` and `reduce` helpers for working with collections (i.e. `Map`, `Set`, `Array`) more easily. Specifically, `map` can map one `Map` to another.

* feat(admin): make write-in transcription work

This simplifies a few things in `WriteInScreen` and `WriteInTranscriptionScreen` around data management by removing the `useEffect` + `setSomeState` pattern. We can just compute things on the fly from the query data and we'll do something to make it faster if it's slow.

Responsibility for the current transcription "page" now belongs to `WriteInTranscriptionScreen`. In changing to react-query and `useWriteInsQuery` we reload the write-in data every time it changes. This isn't really ideal, but will probably work for now. As a result of that the `useEffect` that was trying to reset `paginationIdx` to 0 was being triggered on every transcription. Instead, we now let the transcription screen handle it and use `key=${contestId}` to ensure the new `offset` is reset in the event we somehow switch contests without unmounting the component first.

Transcription is now handled by a react-query mutation, wrapped in `useTranscribeWriteInMutation`. This ensures that the necessary data is reloaded when transcription happens.