Misc. WIA fixes (#2577)
* fix(types): correct the types for `loadWriteInAdjudications`

* fix(admin): use the right `id` for transcribed value input

The `id` of the text input must match the `for` of the label.

* refactor(admin): use filtered write-in adjudications

* fix(admin): clear adjudications alongside CVRs

If the CVR files are deleted then we delete all the related data as well. We forgot to delete the `write_in_adjudications` table entries, which meant they could stick around and affect subsequently-loaded CVR files.

* fix(admin): make adjudicating official candidates work

Closes #2576

We weren't properly passing the `adjudicatedOptionId` value when initially creating an adjudication for an official candidate. We were doing so when changing the adjudication, so that flow actually did work correctly and was the only way to associate a transcription with an official candidate.

* test(admin): add server tests for official candidate adjudication

Provides some insurance against a regression like #2576.