feat(res-to-ts): count `application/xml` as text (#2598)
* fix(scan): print stack trace on `retry-scan` crash

* fix(scan): make `call` throw if the worker threw

This ensures the caller does not inadvertently assume success by casting to the expected type, as `retry-scan` does. Uses a `Result` at the top level instead of an object with an `output` property to distinguish success from failure.

* feat(res-to-ts): count `application/xml` as text

I'll be adding an `.xml` fixture soon and I need to be able to use `.asText()` with it.