fix(scan-nh): parse more styles of constitutional questions (#2625)
* fix(scan-nh): parse more styles of constitutional questions

After receiving XML definitions from NH we found that the previous parsing implementation did not fully parse the HTML gibberish embedded in the `<Questions>` element. This commit takes a more structured approach and splits the function out into its own file to facilitate testing.

* build(admin): stub `jsdom` when used with vite

We don't actually want `jsdom` in the browser because it doesn't actually work there–it requires NodeJS built-in modules. Also updates the stub for `fs` since it was out of date.