feat(scan): add `scale` option to `render-pages` (#2627)
This is useful for generating the images needed for the `ballot-interpreter-nh/bin/convert` binary since it expects 72ppi rather than the default 144ppi that `render-pages` uses.

I changed `writeImageData` to use `pipeline` from `stream/promises` both because it's more idiomatic and because it fixes an issue where the `imageStream` would be done but the `fileWriter` was not yet for PNG images. This wasn't a problem when using the binary for real, but was a problem with tests.