fix(scan-diagnostic): show correct color for expected vs best
I got confused about which is which because our VX ballots actually use `bounds` to mean "the location where I expected the bubble to be" and `scoredOffset` to mean "how far away from `bounds` I had to go to find the best match". The NH data does not do that, and instead the `bounds` is the location of the best match, so we actually need to reverse the `scoredOffset` to work backward to the original guess.