fix(scan-nh): improve bubble matching (#2667)
* feat(image-utils): make `diff` work with grayscale images

* refactor(image-utils): use number aliases for documentation

* fix(scan-nh): align to bubbles & then score marks

Previously we'd start from where we thought the bubble should be and look in an area around that to find the location that gave the best mark score. The problem with this is it disincentivizes actually lining up with the bubble when it's empty or nearly so. Instead, we now compute two scores: one for how well we match the bubble and one for how filled the area is. We take the fill score for the area that had the best match, regardless of whether there might be higher or lower fill scores elsewhere. This means we should get a lot closer to 0% scores for truly empty bubbles and better align on bubbles with faint marks, which may actually reduce their scores a bit. Because of this, I suggest we change our `definite` mark threshold to 8% from 12% (based on extensive manual testing).

* fix(scan-nh): update the oval template

This one is from the July 12 ballot printed by NH scanned using the Plustek scanner at 300dpi and down-sampled.

* fix(scan-nh): only use image-level thresholds

Thesholding only the cropped area sometimes leads to erroneous results. See https://votingworks.slack.com/archives/C02Q3M874SV/p1665599094342529.

* feat(scan-nh): enable debug images for oval interpretation

Enabled with a debug environment variable, e.g. `DEBUG=*:oval-marks:*William-Lovejoy* ./bin/interpret …`.

* ci(scan): bump the timeout

The other changes on this branch may have made things a bit slower.