fix(admin): handle larger CVR file import
CVR files that have a lot of write-in images can be pretty large, i.e. 1GB+. Attempting to read this in as a string is a really bad idea and NodeJS will simply not allow it as it exceeds the maximum length of a string in V8.

This commit fixes the issue by:
a) passing in the file path when importing rather than the contents
b) computing the sha256 hash via a file read stream
c) reading the CVR file one line at a time

We end up reading the file twice (once for the hash, once for the CVR entry import), but overall this is both more efficient and faster even for strings that do fit in memory.