build(all): wait for service build before running
We have a bootstrapping problem. After setting up development and running `pnpm install`, nothing is actually built yet. When running a frontend we run the associated services and their build in watch mode. But because the service has never been built the service fails to run and the build does not complete because `concurrently` kills it when the service run process exits.

This commit fixes the bootstrapping problem by waiting to run the service until the build completes.