fix(admin): clear uploads after request
Previously we were storing uploads in a temporary file in `/tmp`, but they were never getting cleaned up. Eventually the `/tmp` partition would run out of disk space and we'd be unable to upload anything new, and other operations that used `/tmp` would start to fail as well.

Now, we upload to the `uploads` folder within the current workspace. After the request finishes we delete the uploaded file.