fix(scan-nh): enforce configuring test mode
In #2706 I only passed the test mode config through in the VxCentralScan code path. I had added `isTestMode` as a config parameter to the NH `interpret` function, but I'd made it optional so the type checking didn't complain when we didn't pass it in the VxScan code path. This time I made `isTestMode` required and added it in the VxScan code path.