fix(admin): do not duplicate write-ins from multiple CVR files
When importing multiple files that have any duplicate CVRs we store each CVR only once but link them to all the CVR files they are present in via the `cvr_file_entries` table. When getting the write-in records for an election we need to join `write_ins` with `cvr_file_entries` and `cvr_files`, which means that a write-in record will show up however many times its CVR appears in `cvr_file_entries`. To fix this, we `select distinct` to ensure that only one of each write-in is returned.