feat(scan): support loading raw image files (#2755)
* feat(scan): support loading raw image files

Such files are a temporary workaround for the Custom scanner not working with JPEG. This is only a problem because we're using the same approach as `plustekctl` which writes the files to disk, but long term we should consider loading the scanned images directly in memory.

* fix(scan): remove unused `getBallotImageData`

As @jonahkagan pointed out, I copied the check for `qrcode` and the resulting "No QR code found" error but never removed the original location. This removes it and reworks the tests to not need it.