chore(scan): remove adjudication (#2765)
* refactor(scan): add `sheets#finished_adjudication_at`

We never update the front/back values independently anymore, so we'll consolidate the timestamps into one value. First, we add that value before removing the front/back ones.

* chore(scan): remove adjudication workflow remnants

The UI for most of this was already removed some time ago. This removes the underlying API and schema for handling adjudication on VxScan/VxCentralScan. We're going to want to handle all this stuff differently and this is just unneeded clutter.

* test(scan): add tests to bring coverage back up

* refactor(scan): remove unneeded `!!` coercion