refactor(scan): extract CVR export from `Store`
There's a lot of code in exporting CVRs that isn't really the responsibility of `Store`, like resizing images. This commit extracts that functionality to make `Store` leaner and more able to be extracted from `services/scan` for other users like in https://github.com/votingworks/vxsuite/pull/2662.

Additionally, this replaces the pattern of "passing a writable stream" with "returns a readable stream" in the CVR export code. The core functions now just deal with async iterators and use `Readable.from` to convert those to streams, then `pipeline` to pipe those streams to writable streams like HTTP responses or file writers.