feat(scan): save backups directly from the backend
Previously we would stream the data through the client and either download it (Chrome) or write it directly to a file (`kiosk-browser`). Now we directly write to the USB drive, ensuring that we can only write to paths matching `/media/**`.

This also adds the ability to split large files into 4GB chunks when writing to USB drives given the assumption that we're working with FAT32-formatted drives.

The file export functionality, including splitting, lives in a new library in part because `utils` is a large grab bag and adding file system-specific code makes loading `utils` in clients difficult (we have to stub `fs` and related modules).