refactor(vx-scan): Get CVRs for tally from separate endpoint
Instead of using the streaming CVR export endpoint, we add a new endpoint just for loading the CVRs as a single JSON blob (since Grout doesn't support streaming). We don't currently need streaming in this case since images aren't included. Eventually, all of the tally calculation can be moved to the backend, but this is a minimal refactoring step that removes one consumer of the CVR export endpoint. Coming up, I'll refactor the actual exporting of CVRs to USB to happen on the backend.