Ensure good mutation error handling
Our current pattern is to await async function calls in event handlers
without handling errors. This is a bad pattern because any rejected
promises will not be caught and trigger the window uncaught rejection
error.

This wouldn't be the worst thing in the world, except that in Jest
tests, the uncaught rejection error is extremely unreliable and will
often give you an unrelated error message with no information to point
you to the source of the problem.

So I'm recommending a best practice of handling errors at the source.
There are two ways to do this:
1. Use .mutate(), which is fire-and-forget. Errors will bubble up to the error boundary since we have the useErrorBoundary setting set by default in the query client config.
2. Use .mutateAsync(), but wrap in a try/catch.

I suggest we lean towards .mutate(), since that's the default
recommended by react-query. It also is nice because fire-and-forget is
usually how we want to think about event handlers in React.

For some existing code that expected async, I left the mutateAsync there
but made sure to add a try/catch.