feat(codemods): remove everything except `snake-case`
The babel-based codemods were too hard to keep the types working for and we don't actually expect to use them again–they were only there to serve as examples for future codemods. To that end, `ts-morph` is a better utility for making wide ranging changes to TypeScript projects. Thus, this commit keeps one example codemod using `ts-morph` and deletes everything else. It also improves the ergonomics of running it by using `esbuild-runner`.