refactor(types): Separate candidate definition and vote types
We previously used the Candidate type for both the election definition
candidate listings as well as votes (which are only defined in the apps,
not in the election def). These are actually two distinct things - the
definition of a candidate and the recording of a vote. One major
difference is that candidate definitions don't specify write-ins
(because the voter hasn't actually written anything in), whereas votes
must include write-ins as well as candidate votes. Separating them out
into distinct types supports these different use cases better. Also, it
will allow us to more easily change the election definition type.