feat(vx-scan): add Custom scanner support (#2973)
* feat(scan): add `custom-scanner` package

* chore(scan): add `custom-scanner` as a dependency

* refactor(scan): prepare to introduce Custom support

Make it clearer what in VxScan is generic and what is specifically Plustek.

* refactor(scan): move plustek-specific modules & tests

* chore(scan): remove `PLUSTEKCTL_PATH`

We don't use the "fake plustekctl" approach for using the Custom scanner anymore.

* chore(scan): copy the Plustek state machine for Custom

So far there are no changes made to make it work with the Custom scanner; it's just a straight copy of the Plustek one.

* feat(scan): add support for switching scanners

Since both implementations use the Plustek scanner for now there's no functional difference yet.

* refactor(scan): change Custom copy to use Custom names

This still doesn't change the underlying scanner to be the Custom scanner. It just uses "Custom" or "the scanner" instead of "Plustek".

* test(scan): consolidate scanner-agnostic tests

This recombines the scanner-agnostic tests in the duplicated `app_config.test.ts` files into a single `app_config.test.ts`. It also consolidates some of the `createApp` and related helper functionality.

* test(scan): use Plustek code in the Custom tests

This allows us to begin switching the Custom code over to really using `custom-scanner` without breaking all the tests. We can change one test at a time to use the Custom state machine etc.

* feat(scan): begin `custom-scanner` integration

This makes the Custom copy of the state machine use `custom-scanner` and starts moving some tests over to using it.

* feat(scan): migrate some of the tests to use the Custom state machine

* fix(scan): use the scanner-agnostic generic error

* WIP - not all tests are passing in this commit