build(scan-nh): don't build `debug` target, ever
The debug target is quite slow and I don't think we're ever really going to want to build it. This change makes it very hard to accidentally have built it. It also changes the noun used to describe the Rust part of the package from "extension" to the NodeJS-idiomatic "addon".