test(admin): Don't create a default mock API client in renderInAppContext
We need it to be passed in by the test so that the test can call
assertComplete on it. There were a number of mock errors in the test
output that weren't causing the test to fail because of this.

As a result of this change, I had to update a bunch of test files to
create and pass in an API client mock.