test(admin): Use default query client options
I had a hunch that the API mock errors fixed in the previous commit may have been the cause of the triggering of the error boundary. Turning the error boundary back on works, which makes me think that turning it off caused react-query to swallow the errors that were being thrown by the API mock. That swallowing of errors combined with the lack of calling assertComplete made it seem like the tests were working when really they weren't. Lesson learned: if you see the error boundary, that's probably a real issue!