fix(scan-nh): better handle missed bottom marks
When looking for timing marks we sometimes ignore ones because they don't quite fit the shape/size/etc we're looking for. This is typically due to stretching of the ballot image due to a non-uniform scanning rate. If the bottom row of timing marks was missing any then we'd fail to decode the metadata encoded by those marks.

This commit fixes that by checking each location of the inferred timing marks along the bottom to see if it is mostly filled in, whether we initially detected a timing mark there or not.