fix(lint): make VS Code+ESLint integration work (#3133)
I keep getting errors in VS Code with the `no-underscore-dangle` rule. It says the configuration has unknown properties. Newer versions of ESLint introduced new config properties that seem to be showing up in the resolved config, but the version of ESLint the VS Code integration ends up picking is sometimes a version that does not know about those config properties, making ESLint in VS Code stop working. This commit fixes that by pinning to a specific ESLint version.