refactor(basics): introduce `IteratorPlus` and async variant (#3148)
* refactor(basics): introduce `IteratorPlus` and async variant

Replaces the free functions like `map` and `take` that worked with iterators and replaces them with an `iter` function that returns either an `IteratorPlus` or `AsyncIteratorPlus`, which are just `Iterable` and `AsyncIterable` with a bunch of useful methods tacked on. The main difference from before is that instead of doing everything inside out you can chain things to make them more readable (and discoverable):

```ts
// before
Array.from(map(
  take(5, naturals()),
  (n) => n * n
));

// after
naturals()
  .take(5)
  .map((n) => n * n)
  .toArray()
```

* refactor(basics): fix the test label for renamed `drop` → `skip`

* refactor(basics): use generators directly

Instead of preemptively wrapping in `iter`, do it at the call site when needed. The `iter((function *() {})())` syntax is just too weird.